Renamed non-explicit package name testsuite to model.
