+++ src/index.js
@@ -2,4 +2,4 @@
 
 exports.filesearch = require('./filesearch');
 exports.interpreter = require('./interpreter');
-exports.testSuite = require('./testsuite');
+exports.model = require('./model');
